move common logic for table to hook
