+++ src/pages/Watchlist/Screener.js
@@ -1,4 +1,4 @@
-import React, { useState, useEffect, useCallback, useMemo } from 'react'
+import React, { useState, useEffect } from 'react'
 import { useScreenerUrl } from '../../ducks/Watchlists/utils'
 import {
   getProjectsByFunction,
@@ -8,23 +8,16 @@ import TopPanel from '../../ducks/Watchlists/Widgets/TopPanel'
 import AssetsTable from '../../ducks/Watchlists/Widgets/Table'
 import { buildFunction } from '../../ducks/Watchlists/Widgets/Filter/utils'
 import Infographics from './Infographics'
-import {
-  DEFAULT_ORDER_BY,
-  DIRECTIONS
-} from '../../ducks/Watchlists/Widgets/Table/Columns/defaults'
 import { SCREENER } from '../../ducks/Watchlists/detector'
 import { tableQuery } from '../../ducks/Watchlists/gql'
 import { useRecent } from '../../ducks/Watchlists/gql/list/hooks'
-import { getColumns } from '../../ducks/Watchlists/Widgets/Table/Columns/builder'
 import { useUpdateWatchlist } from '../../ducks/Watchlists/gql/list/mutations'
 import {
   DEFAULT_SCREENER_FN,
   DEFAULT_SCREENER_ID
 } from '../../ducks/Screener/utils'
+import { useColumns } from '../../ducks/Watchlists/Widgets/Table/hooks'
 import styles from './Screener.module.scss'
-
-const pageSize = 20
-const EMPTY_ARRAY = []
 
 const Screener = ({
   watchlist,
@@ -35,13 +28,17 @@ const Screener = ({
   id
 }) => {
   useRecent(watchlist, SCREENER)
-  const defaultPagination = { page: 1, pageSize: +pageSize }
-  const [pagination, setPagination] = useState(defaultPagination)
-  const [orderBy, setOrderBy] = useState(DEFAULT_ORDER_BY)
-  const [activeColumnsKeys, setActiveColumnsKeys] = useState(EMPTY_ARRAY)
-  const activeColumns = useMemo(() => getColumns(activeColumnsKeys), [
-    activeColumnsKeys
-  ])
+
+  const {
+    pagination,
+    setPagination,
+    orderBy,
+    setOrderBy,
+    fetchData,
+    activeColumns,
+    setActiveColumnsKeys
+  } = useColumns()
+
   const [updateWatchlist, { loading: isUpdating }] = useUpdateWatchlist(
     SCREENER
   )
@@ -114,28 +111,6 @@ const Screener = ({
       setTableLoading(false)
     )
   }
-
-  const fetchData = useCallback(
-    ({ pageSize, sortBy }) => {
-      const { id, desc } = sortBy[0]
-      const activeColumn = activeColumns.find(column => column.key === id)
-      if (!activeColumn) {
-        setOrderBy(DEFAULT_ORDER_BY)
-      } else {
-        const { timeRange, aggregation } = activeColumn
-        const newDirection = desc ? DIRECTIONS.DESC : DIRECTIONS.ASC
-        setOrderBy({
-          metric: id,
-          aggregation,
-          dynamicTo: 'now',
-          dynamicFrom: timeRange,
-          direction: newDirection
-        })
-      }
-      setPagination({ ...pagination, pageSize: +pageSize })
-    },
-    [activeColumns]
-  )
 
   return (
     <>
