+++ src/pages/Watchlist/Watchlist.js
@@ -4,6 +4,7 @@ import { PROJECT } from '../../ducks/Watchlists/detector'
 import TopPanel from '../../ducks/Watchlists/Widgets/TopPanel'
 import { useScreenerUrl } from '../../ducks/Watchlists/utils'
 import { useRecent } from '../../ducks/Watchlists/gql/list/hooks'
+import { useColumns } from '../../ducks/Watchlists/Widgets/Table/hooks'
 import AssetsTemplates from '../../ducks/Watchlists/Widgets/Table/AssetsTemplates'
 import styles from './Watchlist.module.scss'
 
@@ -14,6 +15,16 @@ const WatchlistPage = ({ location, history, watchlist }) => {
     history,
     defaultParams: { isMovement: true }
   })
+
+  const {
+    pagination,
+    setPagination,
+    orderBy,
+    setOrderBy,
+    fetchData,
+    activeColumns,
+    setActiveColumnsKeys
+  } = useColumns()
 
   return (
     <>
